+++ pkg/multiinstance/manager/id.go
@@ -1,0 +1,6 @@
+package manager
+
+import "github.com/kong/kong-operator/v2/ingress-controller/pkg/manager"
+
+// ID is an alias for manager.ID from the ingress-controller package.
+type ID = manager.ID